remove extension selection feature
