+++ src/options/pages/marketplace/useWizard.ts
@@ -14,7 +14,7 @@ import { isEmpty, mapValues, uniq } from "lodash";
 import { PIXIEBRIX_SERVICE_ID } from "@/services/constants";
 import { Schema } from "@/core";
 import { isPrimitive } from "@/utils";
-import ConfigureBody from "@/options/pages/marketplace/ConfigureBody";
+import ExtensionsBody from "@/options/pages/marketplace/ExtensionsBody";
 import OptionsBody from "@/options/pages/marketplace/OptionsBody";
 import ServicesBody from "@/options/pages/marketplace/ServicesBody";
 import PermissionsBody from "@/options/pages/marketplace/PermissionsBody";
@@ -31,7 +31,7 @@ const STEPS: WizardStep[] = [
     }>,
   },
   { key: "services", label: "Select Integrations", Component: ServicesBody },
-  { key: "review", label: "Select Bricks", Component: ConfigureBody },
+  { key: "review", label: "Extensions Contained", Component: ExtensionsBody },
   { key: "activate", label: "Permissions & URLs", Component: PermissionsBody },
 ];
